Removed the timestamp from the file name
This was done to make sure we don't accidently override wikis with our
custom bash function and move commands, but it populates the permalink,
so I removed it from the name and the bash function will also no longer
append it. Instead we are using `mv -i` to get warning before updating
anything.