(enhance) Change the style of the inline code
The initial inline code style was very non-obvious.
I eventually figured out I could change the syntax highlighting for the
inline code block with the following change.
https://www.janmeppe.com/blog/how-to-change-code-blocks-minimal-mistakes/
https://mmistakes.github.io/minimal-mistakes/docs/stylesheets/
https://github.com/mmistakes/minimal-mistakes/discussions/2808

Basically after the original SCSS is loaded, we make some configuration
changes.
The change exclude this style when the <code> tag is in the header <tag>

ChatGPT helped a little to understand what is going on and come up with
final solution.