(enhance) Next and Previous links within category
The Next and Previous blog link on each blog currently points the post
which is chronologically next and previous across all the categories.
IMO that is not useful. so we will modify it such that we refer to next
and previous blog in current category.

NOTE: This can hamper build time once the site becomes large.

Reference : https://talk.jekyllrb.com/t/how-to-link-to-next-and-previous-posts-for-same-blog-category/629/5