(enhance) Add related notes into the page
This was an experiment.
I created a new archive-single.html include from minimal-mistakes theme
which changes the title of the post to <h3> from <h2>.
Since the page__content style will override the archive__item-excerpt
style for the font-size, we change it here.

I may not keep this experiment around.
I could have just linked to the tag #notes and be done with it, this is
just extra optimisation which doesn't matter and I spent an unnecessary
time on this.