(enhance) Qoute of the day
Finally I did this. The idea is to publish the quote I like on a
repository and then show them randomly on my website.
The seed here is date of the day. which will change every day so I would
be getting a random quote from my repository of quotes everyday.

See this repo for how I am managing the quotes:
https://github.com/svaderia/quote-of-the-day